paper and author batch queries
